+++ app/src/main/AndroidManifest.xml
@@ -19,9 +19,9 @@
         <activity
             android:name=".LoginActivity"
             android:label="@string/app_name" />
-        <activity android:name=".HelpingHandActivity" />
-        <activity android:name=".HelpingHandPostActivity" />
-        <activity android:name=".HelpingHandMyrequestsActivity" />
+        <activity android:name=".activities.helpinghand.HelpingHandActivity" />
+        <activity android:name=".activities.helpinghand.HelpingHandPostActivity" />
+        <activity android:name=".activities.helpinghand.HelpingHandMyrequestsActivity" />
         <activity
             android:name=".LandingPage"
             android:theme="@style/AppTheme.NoActionBar">
