+++ app/src/main/java/com/app/venki/up/activities/helpinghand/FirebaseConstants.java
@@ -1,4 +1,4 @@
-package com.app.venki.up;
+package com.app.venki.up.activities.helpinghand;
 
 /**
  * Created by bala on 8/5/16.
